Adding final merge additions and organizing. Added animation
