+++ src/styling/home.css
@@ -1,0 +1,65 @@
+/* Fade-in animation */
+.fade-in {
+    opacity: 0.5;
+    transform: scale(0.97);
+    animation: fadeInAnim 0.5s ease-in forwards;
+}
+
+@keyframes fadeInAnim {
+    from {
+        opacity: 0.5;
+        transform: scale(0.97);
+    }
+    to {
+        opacity: 1;
+        transform: scale(1);
+    }
+}
+
+/* Fade-out animation */
+.fade-out {
+    opacity: 1;
+    transform: scale(1);
+    animation: fadeOutAnim 0.5s ease-out forwards;
+}
+
+@keyframes fadeOutAnim {
+    from {
+        opacity: 1;
+        transform: scale(1);
+    }
+    to {
+        opacity: 0;
+        transform: scale(1.03);
+    }
+}
+
+/* Fade-in animation for staggered effect */
+.staggered-box {
+    opacity: 0;
+    animation: fadeInStaggered 1s ease-out forwards;
+}
+
+.staggered-box:nth-child(1) {
+    animation-delay: 0.2s;
+}
+.staggered-box:nth-child(2) {
+    animation-delay: 0.4s;
+}
+.staggered-box:nth-child(3) {
+    animation-delay: 0.6s;
+}
+.staggered-box:nth-child(4) {
+    animation-delay: 0.8s;
+}
+
+@keyframes fadeInStaggered {
+    from {
+        opacity: 0;
+        transform: translateY(20px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
